hls/m53683: signal sps bounding box at conformance scale
Previously, it was required to comprehend the source scale factor in
order to apply the sequence bounding box to the output point cloud (or
conversely, to apply the output point cloud to the sequence bounding
box).

This change signals the sequence bounding box in the world-coordinate
system of the decoded output without applying any scaling.

+++ tmc3/ply.cpp
@@ -90,7 +90,7 @@ ply::write(
   const PCCPointSet3& cloud,
   const PropertyNameMap& attributeNames,
   double positionScale,
-  Vec3<int32_t> positionOffset,
+  Vec3<double> positionOffset,
   const std::string& fileName,
   bool asAscii)
 {
